finish up all remaining commands from slash-create to discord.js
