+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@discordjs/rest": "^0.5.0",
     "@prisma/client": "4.0.0",
+    "@upstash/redis": "^1.18.5",
     "axios": "^0.27.2",
     "cat-loggr": "^1.1.0",
     "discord-api-types": "^0.36.1",
