+++ src/index.ts
@@ -3,7 +3,11 @@ import { ApplicationClient } from './lib/client';
 
 dotenv.config();
 
-const client = new ApplicationClient();
-client.migrate();
+const admins = [
+  '183908254210981888', // kharann,
+  '164357764020305920' // Tikka
+];
+
+const client = new ApplicationClient(admins);
 
 client.login(process.env.DISCORD_BOT_TOKEN);
