+++ src/schemas/user.ts
@@ -1,7 +1,7 @@
 import { Rank, Region, Role } from '@prisma/client';
 import { z } from 'zod';
 
-export const SetupSchema = z.object({
+export const SetupCommandInputSchema = z.object({
   ign: z.string(),
   region: z.nativeEnum(Region),
   rank: z.nativeEnum(Rank),
@@ -9,7 +9,11 @@ export const SetupSchema = z.object({
   secondary: z.nativeEnum(Role)
 });
 
-export const MatchSchema = z.object({
+export type SetupInput = z.infer<typeof SetupCommandInputSchema>;
+
+export const MatchCommandInputSchema = z.object({
   match_id: z.number(),
   status: z.enum(['BLUE', 'RED', 'REMAKE'])
 });
+
+export type MatchInput = z.infer<typeof MatchCommandInputSchema>;
