+++ src/services/index.ts
@@ -1,4 +1,5 @@
 import { PrismaClient } from '@prisma/client';
+import { Redis } from '@upstash/redis';
 import { initMatchmakingService } from './matchmaking-service';
 import { initQueueService } from './queue-service';
 import { initScrimRepository } from './repo/scrim-repository';
@@ -9,18 +10,24 @@ import Discord, { GatewayIntentBits } from 'discord.js';
 import { initDiscordService } from './discord-service';
 
 const client = new Discord.Client({
-  intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.DirectMessages]
+  intents: [
+    GatewayIntentBits.Guilds,
+    GatewayIntentBits.GuildMessages,
+    GatewayIntentBits.GuildMembers,
+    GatewayIntentBits.DirectMessages
+  ]
 });
 
 client.login(process.env.DISCORD_BOT_TOKEN).then(() => {});
 
 // Clients
 const prisma = new PrismaClient();
+
 // Repositories
 const userRepo = initUserRepository(prisma);
 const scrimRepo = initScrimRepository(prisma);
 // Services
-export const userService = initUserService(userRepo);
+export const userService = initUserService(prisma);
 const matchmakingService = initMatchmakingService();
 export const discordService = initDiscordService(client);
 export const scrimService = initScrimService(scrimRepo, userRepo, matchmakingService, discordService);
